+++ resources/[dg]/dg-vehicles/typescript/src/client/services/index.ts
@@ -11,3 +11,4 @@ import './vehicleModes';
 import './windowtint';
 import './itemupgrades';
 import './fakeplate';
+import './enginesounds';